% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryUniprotKB.R
\name{availableFeaturesUniprotKB}
\alias{availableFeaturesUniprotKB}
\title{Query available protein features in UniprotKB.}
\usage{
availableFeaturesUniprotKB()
}
\value{
a data.frame.
}
\description{
Query available protein features in UniprotKB.
}
\examples{
head(availableFeaturesUniprotKB(), 10)
}
