% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maser.R
\name{counts}
\alias{counts}
\title{Retrieve raw read counts values from a maser object.}
\usage{
counts(events, type)
}
\arguments{
\item{events}{a maser object.}

\item{type}{a character indicating the splice type. Possible values 
are  \code{c("A3SS", "A5SS", "SE", "RI", "MXE")}.}
}
\value{
a data.frame.
}
\description{
Retrieve raw read counts values from a maser object.
}
\examples{
path <- system.file("extdata", file.path("MATS_output"), package = "maser")
hypoxia <- maser(path, c("Hypoxia 0h", "Hypoxia 24h"))
head(counts(hypoxia, "SE"))
}
