% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maser.R
\name{maser}
\alias{maser}
\title{Create a maser object by importing rMATS splicing events.}
\usage{
maser(path, cond_labels, rtype = "ReadsOnTargetAndJunction")
}
\arguments{
\item{path}{a character specifiying the folder containing rMATS output files.}

\item{cond_labels}{a character vector of length 2 describing labels for 
experimental conditions.}

\item{rtype}{a character indicating the read type. Possible values 
are \code{c("ReadsOnTargetAndJunction", "JunctionCountOnly")}.}
}
\value{
A maser object.
}
\description{
Create a maser object by importing rMATS splicing events.
}
\examples{
path <- system.file("extdata", file.path("MATS_output"), package = "maser")
hypoxia <- maser(path, c("Hypoxia 0h", "Hypoxia 24h"))
}
