% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEvents.R
\name{pca}
\alias{pca}
\title{Prinicipal component analysis of PSI distributions.}
\usage{
pca(events, type = c("A3SS", "A5SS", "SE", "RI", "MXE"))
}
\arguments{
\item{events}{a maser object.}

\item{type}{character indicating splice type. Possible values are
\code{c("A3SS", "A5SS", "SE", "RI", "MXE")}}
}
\value{
a ggplot object.
}
\description{
Prinicipal component analysis of PSI distributions.
}
\examples{
path <- system.file("extdata", file.path("MATS_output"), package = "maser")
hypoxia <- maser(path, c("Hypoxia 0h", "Hypoxia 24h"))
hypoxia_filt <- filterByCoverage(hypoxia, avg_reads = 5)
pca(hypoxia_filt, type = "RI")
}
