% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maser.R
\name{print.maser}
\alias{print.maser}
\title{Print a maser object.}
\usage{
\method{print}{maser}(x, ...)
}
\arguments{
\item{x}{a maser object.}

\item{...}{further arguments.}
}
\description{
Print a maser object.
}
\examples{
path <- system.file("extdata", file.path("MATS_output"), package = "maser")
hypoxia <- maser(path, c("Hypoxia 0h", "Hypoxia 24h"))
print(hypoxia)
}
